+++ src/Cwaff-Guns/MacchiAuto.cs
@@ -72,9 +72,9 @@ public class MacchiAuto : CwaffGun
         private static void ApplyExtraBeamStatusEffects(BeamController beam, GameActor gameActor)
         {
             // ETGModConsole.Log($"applying extra status effects");
-            for (int i = 0; i < beam.projectile.statusEffectsToApply.Count; ++i)
+            foreach (GameActorEffect e in beam.projectile.statusEffectsToApply)
                 if (UnityEngine.Random.value < BraveMathCollege.SliceProbability(beam.statusEffectChance, BraveTime.DeltaTime))
-                    gameActor.ApplyEffect(beam.projectile.statusEffectsToApply[i]);
+                    gameActor.ApplyEffect(e);
         }
     }
 }